POC endpoints implemented with DB.

+++ src/main/java/com/tiy/webapp/ContactRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by Paul Dennis on 1/26/2017.
  */
-public interface ContactRepo extends CrudRepository<UserContact, Integer> {
+public interface ContactRepo extends CrudRepository<UserContact, Long> {
 
 }
